+++ baseline/raif_hack/features.py
@@ -1,0 +1,13 @@
+import pandas as pd
+from raif_hack.utils import UNKNOWN_VALUE
+
+def prepare_categorical(df: pd.DataFrame) -> pd.DataFrame:
+    """
+    Заполняет пропущенные категориальные переменные
+    :param df: dataframe, обучающая выборка
+    :return: dataframe
+    """
+    df_new = df.copy()
+    fillna_cols = ['region','city','street','realty_type']
+    df_new[fillna_cols] = df_new[fillna_cols].fillna(UNKNOWN_VALUE)
+    return df_new
